requirements.txt
# Core Framework - LangGraph and LangChain
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-aws>=0.1.0
langchain-openai>=0.2.0
langchain-community>=0.3.0

# HTTP and Web Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
httpx>=0.27.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# AWS Integration
boto3>=1.34.0
botocore>=1.34.0

# Google APIs
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0

# FHIR Resources
fhir.resources>=7.0.0

# Vector Memory
chromadb>=0.4.0
openai>=1.0.0

# Configuration Management
python-dotenv>=1.0.0

# Data Validation
pydantic>=2.0.0
pydantic-settings>=2.0.0

# API Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
websockets>=12.0

# Streamlit Dashboard
streamlit>=1.28.0
plotly>=5.17.0
pydeck>=0.8.0
folium>=0.14.0
streamlit-folium>=0.15.0

# PDF Generation
reportlab>=4.0.0

# Distance Calculations
haversine>=2.8.0

# Additional Utilities
tenacity>=8.2.0
python-dateutil>=2.8.2
